requirements.txt
boto3==1.24.14
botocore==1.27.14
chevron>=0.14
fastapi>=0.68
fire>=0.4.0
Jinja2>=3.1.2
pydantic==1.9.1
pytest>=7.1.2
python-dotenv==0.20.0
python-multipart>=0.0.5
simplejson>=3.17.6
termcolor>=1
uvicorn>=0.14
WTForms==3.0.1